add SDL console
This adds a fake console rendered and handled by SDL 2, featuring:

- cross-platform high ASCII support
- zoomed-in pixel-perfect font rendering with no extra config needed
- batched rendering with SDL 2's rendering subsystem
- better key handling: single-press Esc, Shift/Ctrl/Alt, numpad+numlock
- accurate ASCII console with FreePascal's video/keyboard driver systems
- blinking cursor and text support

To use, run the game with `-sdlconsole` or set `Graphics = "SDLCONSOLE"`
in `config.lua`.

This bundles a `font_console.png` that uses the Toshiba Satellite 8x14
font from int10h.org, with adjustments to trim it down to 8x13 instead.

Other notes:

- colors are hard-coded
- screen size detection doesn't consider title/task bars, other screens
- no screen reader support since it's not a real console